requirements.txt
transformers==4.46.2
accelerate==1.4.0
huggingface_hub==0.28.1
bitsandbytes==0.45.2
typing_extensions==4.12.2
einops==0.8.1
vector-quantize-pytorch==1.21.8
vocos==0.1.0
librosa==0.10.2.post1
torchvision
dotenv
